requirements.txt
streamlit==1.37.0
google-generativeai==0.3.1
beautifulsoup4==4.12.2
requests==2.32.0
python-dotenv==1.0.0
langchain==0.3.7
langchain-community==0.3.1
langchain-text-splitters==0.0.1
langchain-core==0.1.35
langsmith==0.0.83
chromadb==0.4.18
pandas~=2.2.3
faiss-cpu
numpy==1.26.4
tenacity==8.2.3

pip~=24.3.1
attrs~=24.2.0
distro~=1.9.0
protobuf~=5.29.3
pillow~=10.4.0
filelock~=3.17.0
rsa~=4.9
pyasn1~=0.6.1
grpcio~=1.70.0
rich~=13.9.4
proto-plus~=1.26.0
aiohttp~=3.11.13
pyasn1_modules~=0.4.1
cachetools~=5.5.2
certifi~=2024.8.30
urllib3~=1.26.20
packaging~=23.2
grpcio-status~=1.62.3
typing_extensions~=4.12.2
pydantic~=2.9.2
tqdm~=4.67.1
zstandard~=0.23.0
pyparsing~=3.2.0
mypy~=1.7.1
pydantic_core~=2.23.4
Deprecated~=1.2.18
annotated-types~=0.7.0
idna~=3.10
charset-normalizer~=2.0.12
Jinja2~=3.1.4